rm unused aws_cloudfront_origin_request_policy
Because WAFv2 does not provide a fieldmatch value of `ja4_fingerprint` yet
I was hopeful that Cloudfront might apply this origin request policy
before forwarding the request to WAF for evaluation. If that were the case,
then the WAF rule could be rewritten to rely on the `cloudfront-viewer-ja4-fingerprint`
header value.

Unfortunately, after testing, the origin request policy does not
have the desired effect (likely because WAF evaluates the request before Cloudfront does
and origin-related processing) and the WAF request headers payload lack the
desired fingerprint.